+++ src/sections/collection/Collection.tsx
@@ -1,55 +1,51 @@
+import { useTranslation } from 'react-i18next'
 import { Alert, Col, Row } from '@iqss/dataverse-design-system'
-import { DatasetRepository } from '../../dataset/domain/repositories/DatasetRepository'
-import { DatasetsList } from './datasets-list/DatasetsList'
-import { DatasetsListWithInfiniteScroll } from './datasets-list/DatasetsListWithInfiniteScroll'
-import { BreadcrumbsGenerator } from '../shared/hierarchy/BreadcrumbsGenerator'
-
-import styles from './Collection.module.scss'
-import AddDataActionsButton from '../shared/add-data-actions/AddDataActionsButton'
-import { useSession } from '../session/SessionContext'
-import { useCollection } from './useCollection'
 import { CollectionRepository } from '../../collection/domain/repositories/CollectionRepository'
-import { PageNotFound } from '../page-not-found/PageNotFound'
-import { CollectionSkeleton } from './CollectionSkeleton'
-import { CollectionInfo } from './CollectionInfo'
-import { Trans, useTranslation } from 'react-i18next'
+import { useCollection } from './useCollection'
 import { useScrollTop } from '../../shared/hooks/useScrollTop'
+import { useSession } from '../session/SessionContext'
 import { useGetCollectionUserPermissions } from '../../shared/hooks/useGetCollectionUserPermissions'
+import { type UseCollectionQueryParamsReturnType } from './useGetCollectionQueryParams'
+import { BreadcrumbsGenerator } from '../shared/hierarchy/BreadcrumbsGenerator'
+import AddDataActionsButton from '../shared/add-data-actions/AddDataActionsButton'
+import { CollectionItemsPanel } from './collection-items-panel/CollectionItemsPanel'
+import { CollectionInfo } from './CollectionInfo'
+import { CollectionSkeleton } from './CollectionSkeleton'
+import { PageNotFound } from '../page-not-found/PageNotFound'
+import { CreatedAlert } from './CreatedAlert'
 import { PublishCollectionButton } from './publish-collection/PublishCollectionButton'
+import styles from './Collection.module.scss'
 
 interface CollectionProps {
-  repository: CollectionRepository
-  datasetRepository: DatasetRepository
-  id: string
+  collectionRepository: CollectionRepository
+  collectionId: string
   created: boolean
   published: boolean
-  page?: number
+  collectionQueryParams: UseCollectionQueryParamsReturnType
   infiniteScrollEnabled?: boolean
 }
 
 export function Collection({
-  repository,
-  id,
-  datasetRepository,
+  collectionId,
+  collectionRepository,
   created,
   published,
-  page,
-  infiniteScrollEnabled = false
+  collectionQueryParams
 }: CollectionProps) {
+  useTranslation('collection')
   useScrollTop()
   const { user } = useSession()
-  const { collection, isLoading } = useCollection(repository, id, published)
+  const { collection, isLoading } = useCollection(collectionRepository, collectionId, published)
   const { collectionUserPermissions } = useGetCollectionUserPermissions({
-    collectionIdOrAlias: id,
-    collectionRepository: repository
+    collectionIdOrAlias: collectionId,
+    collectionRepository
   })
 
   const canUserAddCollection = Boolean(collectionUserPermissions?.canAddCollection)
   const canUserAddDataset = Boolean(collectionUserPermissions?.canAddDataset)
   const canUserPublishCollection = user && Boolean(collectionUserPermissions?.canPublishCollection)
 
-  const showAddDataActions = user && (canUserAddCollection || canUserAddDataset)
-
+  const showAddDataActions = Boolean(user && (canUserAddCollection || canUserAddDataset))
   const { t } = useTranslation('collection')
 
   if (!isLoading && !collection) {
@@ -65,23 +61,7 @@ export function Collection({
           <>
             <BreadcrumbsGenerator hierarchy={collection.hierarchy} />
             <CollectionInfo collection={collection} />
-            {created && (
-              <Alert variant="success" dismissible={false}>
-                <Trans
-                  t={t}
-                  i18nKey="createdAlert"
-                  components={{
-                    anchor: (
-                      <a
-                        href="https://guides.dataverse.org/en/latest/user/dataverse-management.html"
-                        target="_blank"
-                        rel="noreferrer"
-                      />
-                    )
-                  }}
-                />
-              </Alert>
-            )}
+            {created && <CreatedAlert />}
             {published && (
               <Alert variant="success" dismissible={false}>
                 {t('publishedAlert')}
@@ -89,34 +69,29 @@ export function Collection({
             )}
             {!collection.isReleased && canUserPublishCollection && (
               <div className={styles.container}>
-                <PublishCollectionButton repository={repository} collectionId={collection.id} />
-              </div>
-            )}
-            {showAddDataActions && (
-              <div className={styles.container}>
-                <AddDataActionsButton
-                  collectionId={id}
-                  canAddCollection={canUserAddCollection}
-                  canAddDataset={canUserAddDataset}
+                <PublishCollectionButton
+                  repository={collectionRepository}
+                  collectionId={collection.id}
                 />
               </div>
             )}
           </>
         )}
-        {infiniteScrollEnabled ? (
-          <DatasetsListWithInfiniteScroll
-            datasetRepository={datasetRepository}
-            collectionId={id}
-            key={id}
-          />
-        ) : (
-          <DatasetsList
-            datasetRepository={datasetRepository}
-            page={page}
-            collectionId={id}
-            key={id}
-          />
-        )}
+        <CollectionItemsPanel
+          key={collectionId}
+          collectionId={collectionId}
+          collectionRepository={collectionRepository}
+          collectionQueryParams={collectionQueryParams}
+          addDataSlot={
+            showAddDataActions ? (
+              <AddDataActionsButton
+                collectionId={collectionId}
+                canAddCollection={canUserAddCollection}
+                canAddDataset={canUserAddDataset}
+              />
+            ) : null
+          }
+        />
       </Col>
     </Row>
   )