+++ src/sections/collection/collection-items-panel/items-list/dataset-card/DatasetCard.tsx
@@ -1,0 +1,32 @@
+import { DatasetItemTypePreview } from '@/dataset/domain/models/DatasetItemTypePreview'
+import { DatasetCardHeader } from './DatasetCardHeader'
+import { DatasetCardThumbnail } from './DatasetCardThumbnail'
+import { DatasetCardInfo } from './DatasetCardInfo'
+import styles from './DatasetCard.module.scss'
+
+interface DatasetCardProps {
+  datasetPreview: DatasetItemTypePreview
+}
+
+export function DatasetCard({ datasetPreview }: DatasetCardProps) {
+  return (
+    <article className={styles['card-main-container']} data-testid="dataset-card">
+      <DatasetCardHeader
+        persistentId={datasetPreview.persistentId}
+        version={datasetPreview.version}
+      />
+      <div className={styles['thumbnail-and-info-wrapper']}>
+        <DatasetCardThumbnail
+          persistentId={datasetPreview.persistentId}
+          version={datasetPreview.version}
+          thumbnail={datasetPreview.thumbnail}
+        />
+        <DatasetCardInfo
+          version={datasetPreview.version}
+          releaseOrCreateDate={datasetPreview.releaseOrCreateDate}
+          description={datasetPreview.description}
+        />
+      </div>
+    </article>
+  )
+}