+++ src/stories/collection/Collection.stories.tsx
@@ -2,13 +2,9 @@ import type { Meta, StoryObj } from '@storybook/react'
 import { Collection } from '../../sections/collection/Collection'
 import { WithI18next } from '../WithI18next'
 import { WithLayout } from '../WithLayout'
-import { DatasetMockRepository } from '../dataset/DatasetMockRepository'
-import { DatasetLoadingMockRepository } from '../dataset/DatasetLoadingMockRepository'
-import { NoDatasetsMockRepository } from '../dataset/NoDatasetsMockRepository'
 import { WithLoggedInUser } from '../WithLoggedInUser'
 import { CollectionMockRepository } from './CollectionMockRepository'
 import { CollectionLoadingMockRepository } from './CollectionLoadingMockRepository'
-import { NoCollectionMockRepository } from './NoCollectionMockRepository'
 
 const meta: Meta<typeof Collection> = {
   title: 'Pages/Collection',
@@ -26,24 +22,15 @@ type Story = StoryObj<typeof Collection>
 export const Default: Story = {
   render: () => (
     <Collection
-      repository={new CollectionMockRepository()}
-      datasetRepository={new DatasetMockRepository()}
-      id="collection"
+      collectionRepository={new CollectionMockRepository()}
+      collectionId="collection"
       created={false}
       published={false}
-    />
-  )
-}
-
-export const InfiniteScrollingEnabled: Story = {
-  render: () => (
-    <Collection
-      repository={new CollectionMockRepository()}
-      datasetRepository={new DatasetMockRepository()}
-      id="collection"
-      infiniteScrollEnabled={true}
-      created={false}
-      published={false}
+      collectionQueryParams={{
+        pageQuery: 1,
+        searchQuery: undefined,
+        typesQuery: undefined
+      }}
     />
   )
 }
@@ -51,23 +38,11 @@ export const InfiniteScrollingEnabled: Story = {
 export const Loading: Story = {
   render: () => (
     <Collection
-      repository={new CollectionLoadingMockRepository()}
-      datasetRepository={new DatasetLoadingMockRepository()}
-      id="collection"
+      collectionRepository={new CollectionLoadingMockRepository()}
+      collectionId="collection"
       created={false}
       published={false}
-    />
-  )
-}
-
-export const NoResults: Story = {
-  render: () => (
-    <Collection
-      repository={new NoCollectionMockRepository()}
-      datasetRepository={new NoDatasetsMockRepository()}
-      id="collection"
-      created={false}
-      published={false}
+      collectionQueryParams={{ pageQuery: 1, searchQuery: undefined, typesQuery: undefined }}
     />
   )
 }
@@ -76,11 +51,11 @@ export const LoggedIn: Story = {
   decorators: [WithLoggedInUser],
   render: () => (
     <Collection
-      repository={new CollectionMockRepository()}
-      datasetRepository={new DatasetMockRepository()}
-      id="collection"
+      collectionRepository={new CollectionMockRepository()}
+      collectionId="collection"
       created={false}
       published={false}
+      collectionQueryParams={{ pageQuery: 1, searchQuery: undefined, typesQuery: undefined }}
     />
   )
 }
@@ -89,11 +64,11 @@ export const Created: Story = {
   decorators: [WithLoggedInUser],
   render: () => (
     <Collection
-      repository={new CollectionMockRepository()}
-      datasetRepository={new NoDatasetsMockRepository()}
-      id="collection"
+      collectionRepository={new CollectionMockRepository()}
+      collectionId="collection"
       created={true}
       published={false}
+      collectionQueryParams={{ pageQuery: 1, searchQuery: undefined, typesQuery: undefined }}
     />
   )
 }
@@ -101,11 +76,11 @@ export const Published: Story = {
   decorators: [WithLoggedInUser],
   render: () => (
     <Collection
-      repository={new CollectionMockRepository()}
-      datasetRepository={new NoDatasetsMockRepository()}
-      id="collection"
+      collectionRepository={new CollectionMockRepository()}
+      collectionId="collection"
       created={false}
       published={true}
+      collectionQueryParams={{ pageQuery: 1, searchQuery: undefined, typesQuery: undefined }}
     />
   )
 }